Renamed 'cpf.logger_functions' to 'cpf.logging'
