+++ src/cpf/IO_functions.py
@@ -15,7 +15,7 @@ from pathlib import Path
 
 import numpy as np
 
-from cpf.logger_functions import CPFLogger
+from cpf.logging import CPFLogger
 
 logger = CPFLogger("cpf.IO_functions")
