+++ src/cpf/lmfit_model.py
@@ -32,7 +32,7 @@ from lmfit import Model, Parameters
 import cpf.peak_functions as pf
 import cpf.series_constraints as sc
 import cpf.series_functions as sf
-from cpf.logger_functions import CPFLogger
+from cpf.logging import CPFLogger
 
 logger = CPFLogger("cpf.lmfit_model")
